requirements.txt
pyTelegramBotAPI~=4.7.0
mysql-connector-python~=8.0.30
requests~=2.28.1
vk-api~=11.9.9